+++ client/src/components/detailsOfCards-layout/flippableCardContainer.tsx
@@ -1,4 +1,4 @@
-import "./styles/flippableCardContainer.css";
+import "../styles/flippableCardContainer.css";
 import { SingleCard } from "./SingleCard";
 import { CSSTransition } from "react-transition-group";
 import { useState } from "react";
